+++ src/listeners/panel/DataListener.ts
@@ -232,6 +232,8 @@ export class DataListener extends BaseListener {
       return;
     }
 
+    const titleField = (Settings.get(SETTING_SEO_TITLE_FIELD) as string) || DefaultFields.Title;
+
     const editor = window.activeTextEditor;
 
     let article;
@@ -248,6 +250,10 @@ export class DataListener extends BaseListener {
 
     const contentType = ArticleHelper.getContentType(article.data);
 
+    if (!value && field !== titleField && contentType.clearEmpty) {
+      value = undefined;
+    }
+
     const dateFields = ContentType.findFieldsByTypeDeep(contentType.fields, 'datetime');
     const imageFields = ContentType.findFieldsByTypeDeep(contentType.fields, 'image');
     const fileFields = ContentType.findFieldsByTypeDeep(contentType.fields, 'file');
@@ -263,6 +269,7 @@ export class DataListener extends BaseListener {
       }
     });
 
+    // Check multi-image fields
     const multiImageFieldsArray = imageFields.find((f: Field[]) => {
       const lastField = f?.[f.length - 1];
       if (lastField) {
@@ -270,6 +277,7 @@ export class DataListener extends BaseListener {
       }
     });
 
+    // Check multi-file fields
     const multiFileFieldsArray = fileFields.find((f: Field[]) => {
       const lastField = f?.[f.length - 1];
       if (lastField) {
@@ -277,6 +285,7 @@ export class DataListener extends BaseListener {
       }
     });
 
+    // Check date fields
     if (dateFieldsArray && dateFieldsArray.length > 0) {
       for (const dateField of dateFieldsArray) {
         if (field === dateField.name && value) {
@@ -320,6 +329,18 @@ export class DataListener extends BaseListener {
       } else {
         parentObj[field] = value;
       }
+    }
+
+    // Clear the field if it is empty
+    if (
+      value === undefined ||
+      (value instanceof Array && value.length === 0 && contentType.clearEmpty)
+    ) {
+      delete parentObj[field];
+    }
+
+    if (Object.keys(parentObj).length === 0 && field !== titleField && contentType.clearEmpty) {
+      delete article.data[parents![0]];
     }
 
     if (editor) {